* Moves GSendBug, GSendAux and initialisation of easylogging++ into the Server library only so it can be used for dynamic loading.
* Preparation for removal of autotools

+++ SOURCE/Server/AuctionHouse.h
@@ -86,8 +86,8 @@ public:
 	unsigned long mBuyPriceCopperEnd;
 	unsigned long mBuyPriceCreditsStart;
 	unsigned long mBuyPriceCreditsEnd;
-	unsigned long mLevelStart;
-	unsigned long mLevelEnd;
+	short mLevelStart;
+	short mLevelEnd;
 
 	std::string mSearch;
 };
